requirements.txt
pystray>=0.19.0
Pillow>=10.0.0
PyYAML>=6.0
pyinstaller>=6.0.0

# SAP connectivity via JCo + JPype (requires sapjco3.jar + native lib from SAP)
jpype1>=1.5.0
# pyrfc>=3.0.0  # DEPRECATED — replaced by JCo/JPype

# Database drivers (install as needed per target DB)
# hdbcli>=2.0.0          # SAP HANA
# oracledb>=2.0.0        # Oracle
# pyodbc>=5.0.0          # Microsoft SQL Server
# psycopg2-binary>=2.9   # PostgreSQL
